+++ src/config.rs
@@ -1,6 +1,7 @@
 use crate::compressor;
 use crate::filters;
 use crate::mixer;
+use crate::wavtools::{find_data_in_wav_stream, WavParams};
 use parking_lot::RwLock;
 use serde::{de, Deserialize, Serialize};
 //use serde_with;
@@ -16,6 +17,7 @@ use std::path::{Path, PathBuf};
 use crate::PrcFmt;
 type Res<T> = Result<T, Box<dyn error::Error>>;
 
+#[derive(Clone)]
 pub struct Overrides {
     pub samplerate: Option<usize>,
     pub sample_format: Option<SampleFormat>,
@@ -119,7 +121,67 @@ impl fmt::Display for SampleFormat {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[allow(clippy::upper_case_acronyms)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
+// Similar to SampleFormat, but also includes TEXT
+pub enum FileFormat {
+    TEXT,
+    S16LE,
+    S24LE,
+    S24LE3,
+    S32LE,
+    FLOAT32LE,
+    FLOAT64LE,
+}
+
+impl FileFormat {
+    pub fn bits_per_sample(&self) -> usize {
+        match self {
+            FileFormat::S16LE => 16,
+            FileFormat::S24LE => 24,
+            FileFormat::S24LE3 => 24,
+            FileFormat::S32LE => 32,
+            FileFormat::FLOAT32LE => 32,
+            FileFormat::FLOAT64LE => 64,
+            FileFormat::TEXT => 0,
+        }
+    }
+
+    pub fn bytes_per_sample(&self) -> usize {
+        match self {
+            FileFormat::S16LE => 2,
+            FileFormat::S24LE => 4,
+            FileFormat::S24LE3 => 3,
+            FileFormat::S32LE => 4,
+            FileFormat::FLOAT32LE => 4,
+            FileFormat::FLOAT64LE => 8,
+            FileFormat::TEXT => 0,
+        }
+    }
+
+    pub fn from_sample_format(sample_format: &SampleFormat) -> Self {
+        match sample_format {
+            SampleFormat::S16LE => FileFormat::S16LE,
+            SampleFormat::S24LE => FileFormat::S24LE,
+            SampleFormat::S24LE3 => FileFormat::S24LE3,
+            SampleFormat::S32LE => FileFormat::S32LE,
+            SampleFormat::FLOAT32LE => FileFormat::FLOAT32LE,
+            SampleFormat::FLOAT64LE => FileFormat::FLOAT64LE,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+#[serde(tag = "type")]
+pub enum Signal {
+    Sine { freq: f64, level: PrcFmt },
+    Square { freq: f64, level: PrcFmt },
+    WhiteNoise { level: PrcFmt },
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 #[serde(tag = "type")]
 pub enum CaptureDevice {
@@ -142,8 +204,8 @@ pub enum CaptureDevice {
         device: String,
         format: SampleFormat,
     },
-    #[serde(alias = "FILE", alias = "file")]
-    File(CaptureDeviceFile),
+    RawFile(CaptureDeviceRawFile),
+    WavFile(CaptureDeviceWavFile),
     #[serde(alias = "STDIN", alias = "stdin")]
     Stdin(CaptureDeviceStdin),
     #[cfg(target_os = "macos")]
@@ -168,6 +230,11 @@ pub enum CaptureDevice {
         channels: usize,
         device: String,
     },
+    SignalGenerator {
+        #[serde(deserialize_with = "validate_nonzero_usize")]
+        channels: usize,
+        signal: Signal,
+    },
 }
 
 impl CaptureDevice {
@@ -179,7 +246,10 @@ impl CaptureDevice {
             CaptureDevice::Bluez(dev) => dev.channels,
             #[cfg(feature = "pulse-backend")]
             CaptureDevice::Pulse { channels, .. } => *channels,
-            CaptureDevice::File(dev) => dev.channels,
+            CaptureDevice::RawFile(dev) => dev.channels,
+            CaptureDevice::WavFile(dev) => {
+                dev.wav_info().map(|info| info.channels).unwrap_or_default()
+            }
             CaptureDevice::Stdin(dev) => dev.channels,
             #[cfg(target_os = "macos")]
             CaptureDevice::CoreAudio(dev) => dev.channels,
@@ -196,13 +266,14 @@ impl CaptureDevice {
                 )
             ))]
             CaptureDevice::Jack { channels, .. } => *channels,
+            CaptureDevice::SignalGenerator { channels, .. } => *channels,
         }
     }
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
 #[serde(deny_unknown_fields)]
-pub struct CaptureDeviceFile {
+pub struct CaptureDeviceRawFile {
     #[serde(deserialize_with = "validate_nonzero_usize")]
     pub channels: usize,
     pub filename: String,
@@ -215,7 +286,7 @@ pub struct CaptureDeviceFile {
     pub read_bytes: Option<usize>,
 }
 
-impl CaptureDeviceFile {
+impl CaptureDeviceRawFile {
     pub fn extra_samples(&self) -> usize {
         self.extra_samples.unwrap_or_default()
     }
@@ -224,6 +295,33 @@ impl CaptureDeviceFile {
     }
     pub fn read_bytes(&self) -> usize {
         self.read_bytes.unwrap_or_default()
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct CaptureDeviceWavFile {
+    pub filename: String,
+    #[serde(default)]
+    pub extra_samples: Option<usize>,
+}
+
+impl CaptureDeviceWavFile {
+    pub fn extra_samples(&self) -> usize {
+        self.extra_samples.unwrap_or_default()
+    }
+
+    pub fn wav_info(&self) -> Res<WavParams> {
+        let fname = &self.filename;
+        let f = match File::open(fname) {
+            Ok(f) => f,
+            Err(err) => {
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
+                return Err(ConfigError::new(&msg).into());
+            }
+        };
+        let file = BufReader::new(&f);
+        find_data_in_wav_stream(file)
     }
 }
 
@@ -336,12 +434,16 @@ pub enum PlaybackDevice {
         channels: usize,
         filename: String,
         format: SampleFormat,
+        #[serde(default)]
+        wav_header: Option<bool>,
     },
     #[serde(alias = "STDOUT", alias = "stdout")]
     Stdout {
         #[serde(deserialize_with = "validate_nonzero_usize")]
         channels: usize,
         format: SampleFormat,
+        #[serde(default)]
+        wav_header: Option<bool>,
     },
     #[cfg(target_os = "macos")]
     #[serde(alias = "COREAUDIO", alias = "coreaudio")]
@@ -624,45 +726,6 @@ pub enum Filter {
         description: Option<String>,
         parameters: LimiterParameters,
     },
-}
-
-#[allow(clippy::upper_case_acronyms)]
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub enum FileFormat {
-    TEXT,
-    S16LE,
-    S24LE,
-    S24LE3,
-    S32LE,
-    FLOAT32LE,
-    FLOAT64LE,
-}
-
-impl FileFormat {
-    pub fn bits_per_sample(&self) -> usize {
-        match self {
-            FileFormat::S16LE => 16,
-            FileFormat::S24LE => 24,
-            FileFormat::S24LE3 => 24,
-            FileFormat::S32LE => 32,
-            FileFormat::FLOAT32LE => 32,
-            FileFormat::FLOAT64LE => 64,
-            FileFormat::TEXT => 0,
-        }
-    }
-
-    pub fn bytes_per_sample(&self) -> usize {
-        match self {
-            FileFormat::S16LE => 2,
-            FileFormat::S24LE => 4,
-            FileFormat::S24LE3 => 3,
-            FileFormat::S32LE => 4,
-            FileFormat::FLOAT32LE => 4,
-            FileFormat::FLOAT64LE => 8,
-            FileFormat::TEXT => 0,
-        }
-    }
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
@@ -1218,7 +1281,8 @@ impl PipelineStepMixer {
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
 pub struct PipelineStepFilter {
-    pub channel: usize,
+    #[serde(default)]
+    pub channels: Option<Vec<usize>>,
     pub names: Vec<String>,
     #[serde(default)]
     pub description: Option<String>,
@@ -1283,7 +1347,7 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
     let file = match File::open(filename) {
         Ok(f) => f,
         Err(err) => {
-            let msg = format!("Could not open config file '{filename}'. Error: {err}");
+            let msg = format!("Could not open config file '{filename}'. Reason: {err}");
             return Err(ConfigError::new(&msg).into());
         }
     };
@@ -1292,7 +1356,7 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
     let _number_of_bytes: usize = match buffered_reader.read_to_string(&mut contents) {
         Ok(number_of_bytes) => number_of_bytes,
         Err(err) => {
-            let msg = format!("Could not read config file '{filename}'. Error: {err}");
+            let msg = format!("Could not read config file '{filename}'. Reason: {err}");
             return Err(ConfigError::new(&msg).into());
         }
     };
@@ -1303,15 +1367,24 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
             return Err(ConfigError::new(&msg).into());
         }
     };
-    //Ok(configuration)
-    //apply_overrides(&mut configuration);
-    //replace_tokens_in_config(&mut configuration);
-    //replace_relative_paths_in_config(&mut configuration, filename);
     Ok(configuration)
 }
 
 fn apply_overrides(configuration: &mut Configuration) {
-    let overrides = OVERRIDES.read();
+    let mut overrides = OVERRIDES.read().clone();
+    // Only one match arm for now, might be more later.
+    #[allow(clippy::single_match)]
+    match &configuration.devices.capture {
+        CaptureDevice::WavFile(dev) => {
+            if let Ok(wav_info) = dev.wav_info() {
+                overrides.channels = Some(wav_info.channels);
+                overrides.sample_format = Some(wav_info.sample_format);
+                overrides.samplerate = Some(wav_info.sample_rate);
+                debug!("Updating overrides with values from wav input file, rate {}, format: {}, channels: {}", wav_info.sample_rate, wav_info.sample_format, wav_info.channels);
+            }
+        }
+        _ => {}
+    }
     if let Some(rate) = overrides.samplerate {
         let cfg_rate = configuration.devices.samplerate;
         let cfg_chunksize = configuration.devices.chunksize;
@@ -1331,7 +1404,7 @@ fn apply_overrides(configuration: &mut Configuration) {
             configuration.devices.chunksize = scaled_chunksize;
             #[allow(unreachable_patterns)]
             match &mut configuration.devices.capture {
-                CaptureDevice::File(dev) => {
+                CaptureDevice::RawFile(dev) => {
                     let new_extra = dev.extra_samples() * rate / cfg_rate;
                     debug!(
                         "Scale extra samples: {} -> {}",
@@ -1364,7 +1437,7 @@ fn apply_overrides(configuration: &mut Configuration) {
         debug!("Apply override for extra_samples: {}", extra);
         #[allow(unreachable_patterns)]
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
+            CaptureDevice::RawFile(dev) => {
                 dev.extra_samples = Some(extra);
             }
             CaptureDevice::Stdin(dev) => {
@@ -1376,9 +1449,10 @@ fn apply_overrides(configuration: &mut Configuration) {
     if let Some(chans) = overrides.channels {
         debug!("Apply override for capture channels: {}", chans);
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
+            CaptureDevice::RawFile(dev) => {
                 dev.channels = chans;
             }
+            CaptureDevice::WavFile(_dev) => {}
             CaptureDevice::Stdin(dev) => {
                 dev.channels = chans;
             }
@@ -1415,14 +1489,18 @@ fn apply_overrides(configuration: &mut Configuration) {
             CaptureDevice::Jack { channels, .. } => {
                 *channels = chans;
             }
+            CaptureDevice::SignalGenerator { channels, .. } => {
+                *channels = chans;
+            }
         }
     }
     if let Some(fmt) = overrides.sample_format {
         debug!("Apply override for capture sample format: {}", fmt);
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
+            CaptureDevice::RawFile(dev) => {
                 dev.format = fmt;
             }
+            CaptureDevice::WavFile(_dev) => {}
             CaptureDevice::Stdin(dev) => {
                 dev.format = fmt;
             }
@@ -1459,6 +1537,7 @@ fn apply_overrides(configuration: &mut Configuration) {
             CaptureDevice::Jack { .. } => {
                 error!("Not possible to override capture format for Jack, ignoring");
             }
+            CaptureDevice::SignalGenerator { .. } => {}
         }
     }
 }
@@ -1767,6 +1846,31 @@ pub fn validate_config(conf: &mut Configuration, filename: Option<&str>) -> Res<
             .into());
         }
     }
+    if let CaptureDevice::RawFile(dev) = &conf.devices.capture {
+        let fname = &dev.filename;
+        match File::open(fname) {
+            Ok(f) => f,
+            Err(err) => {
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
+                return Err(ConfigError::new(&msg).into());
+            }
+        };
+    }
+    if let CaptureDevice::WavFile(dev) = &conf.devices.capture {
+        let fname = &dev.filename;
+        let f = match File::open(fname) {
+            Ok(f) => f,
+            Err(err) => {
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
+                return Err(ConfigError::new(&msg).into());
+            }
+        };
+        let file = BufReader::new(&f);
+        let _wav_info = find_data_in_wav_stream(file).map_err(|err| {
+            let msg = format!("Error reading wav file '{fname}'. Reason: {err}");
+            ConfigError::new(&msg)
+        })?;
+    }
     let mut num_channels = conf.devices.capture.channels();
     let fs = conf.devices.samplerate;
     if let Some(pipeline) = &conf.pipeline {
@@ -1807,9 +1911,20 @@ pub fn validate_config(conf: &mut Configuration, filename: Option<&str>) -> Res<
                 }
                 PipelineStep::Filter(step) => {
                     if !step.is_bypassed() {
-                        if step.channel >= num_channels {
-                            let msg = format!("Use of non existing channel {}", step.channel);
-                            return Err(ConfigError::new(&msg).into());
+                        if let Some(channels) = &step.channels {
+                            for channel in channels {
+                                if *channel >= num_channels {
+                                    let msg = format!("Use of non existing channel {}", channel);
+                                    return Err(ConfigError::new(&msg).into());
+                                }
+                            }
+                            for idx in 1..channels.len() {
+                                if channels[idx..].contains(&channels[idx - 1]) {
+                                    let msg =
+                                        format!("Use of duplicated channel {}", &channels[idx - 1]);
+                                    return Err(ConfigError::new(&msg).into());
+                                }
+                            }
                         }
                         for name in &step.names {
                             if let Some(filters) = &conf.filters {